requirements.txt
google-cloud-storage==1.28.1
python-dotenv>=0.20.0
fastapi>=0.85.0
uvicorn>=0.18.3
PyJWT>=2.4.0
tinydb>=4.7.0
bcrypt